+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - utilities:
   - HKT1 (`F<_>`)
+  - EqT: To assert equality of two types
 - algebraic structures:
   - Magma, Semigroup, Monoid, Group
   - MagmaK, SemigroupK, MonoidK, GroupK, Functor (for `F<_>`)
+  - Magmaal, Semigroupal, Monoidal, Apply, Applicative, FlatMap, Monad
 - algebraic properties:
   - Totality, Associativity, Identity, Inverse, Commutativity